+++ pycommon_database/audit_mongo.py
@@ -15,6 +15,7 @@ class Action(enum.IntEnum):
     Insert = 1
     Update = 2
     Delete = 3
+    Rollback = 4
 
 
 def _create_from(model, base):
@@ -26,11 +27,15 @@ def _common_audit(model, base):
         """
         Class providing Audit fields for a MONGODB model.
         """
-        _model = model
+        revision = Column(int, is_primary_key=True, index_type=IndexType.Unique)
 
-        audit_user = Column(str, is_primary_key=True)
-        audit_date_utc = Column(datetime.datetime, is_primary_key=True, index_type=IndexType.Unique)
+        audit_user = Column(str)
+        audit_date_utc = Column(datetime.datetime)
         audit_action = Column(Action)
+
+        @classmethod
+        def get_response_model(cls, namespace):
+            return namespace.model('Audit' + model.__name__, cls._flask_restplus_fields(namespace))
 
         @classmethod
         def audit_add(cls, model_as_dict: dict):
@@ -51,11 +56,12 @@ def _common_audit(model, base):
             """
             :param model_to_query: Arguments that can directly be provided to Mongo.
             """
-            for removed_dict_model in cls._model.__collection__.find(model_to_query):
+            for removed_dict_model in model.__collection__.find(model_to_query):
                 cls._audit_action(action=Action.Delete, model_as_dict=removed_dict_model)
 
         @classmethod
         def _audit_action(cls, action: Action, model_as_dict: dict):
+            model_as_dict[cls.revision.name] = cls._increment(cls.revision.name)
             model_as_dict[cls.audit_user.name] = current_user_name()
             model_as_dict[cls.audit_date_utc.name] = datetime.datetime.utcnow()
             model_as_dict[cls.audit_action.name] = action.value
@@ -70,12 +76,29 @@ def _versioning_audit(model, base):
         """
         Class providing the audit for all versioned MONGODB models.
         """
-        table_name = Column(str, is_primary_key=True)
+        table_name = Column(str, is_primary_key=True, index_type=IndexType.Unique)
         revision = Column(int, is_primary_key=True, index_type=IndexType.Unique)
 
-        audit_user = Column(str, is_primary_key=True)
-        audit_date_utc = Column(datetime.datetime, is_primary_key=True, index_type=IndexType.Unique)
+        audit_user = Column(str)
+        audit_date_utc = Column(datetime.datetime)
         audit_action = Column(Action)
+
+        @classmethod
+        def query_get_parser(cls):
+            query_get_parser = super().query_get_parser()
+            query_get_parser.remove_argument(cls.table_name.name)
+            return query_get_parser
+
+        @classmethod
+        def get_response_model(cls, namespace):
+            all_fields = cls._flask_restplus_fields(namespace)
+            del all_fields[cls.table_name.name]
+            return namespace.model('AuditModel', all_fields)
+
+        @classmethod
+        def get_all(cls, **model_to_query):
+            model_to_query[cls.table_name.name] = model.__tablename__
+            return super().get_all(**model_to_query)
 
         @classmethod
         def audit_add(cls, revision: int):
@@ -90,9 +113,13 @@ def _versioning_audit(model, base):
             cls._audit_action(action=Action.Delete, revision=revision)
 
         @classmethod
+        def audit_rollback(cls, revision: int):
+            cls._audit_action(action=Action.Rollback, revision=revision)
+
+        @classmethod
         def _audit_action(cls, action: Action, revision: int):
             cls.__collection__.insert_one({
-                cls.table_name: model.__tablename__,
+                cls.table_name.name: model.__tablename__,
                 cls.revision.name: revision,
                 cls.audit_user.name: current_user_name(),
                 cls.audit_date_utc.name: datetime.datetime.utcnow(),