+++ pycommon_database/database_sqlalchemy.py
@@ -318,11 +318,6 @@ class CRUDModel:
         return namespace.model(cls.__name__, cls._flask_restplus_fields())
 
     @classmethod
-    def get_audit_response_model(cls, namespace):
-        if cls.audit_model:
-            return namespace.model('Audit' + cls.__name__, cls.audit_model._flask_restplus_fields())
-
-    @classmethod
     def _flask_restplus_fields(cls):
         return {
             marshmallow_field.name: _get_rest_plus_type(marshmallow_field)(